general create_retrying func for all solvers
